+++ lib/grinder_tools.dart
@@ -61,6 +61,7 @@ void defaultClean([GrinderContext context]) => delete(buildDir);
 /**
  * A utility class to run tests for your project.
  */
+@Deprecated('see [TestRunner]')
 class Tests {
   /**
    * Run command-line tests. You can specify the base directory (`test`), and
@@ -172,6 +173,73 @@ class Tests {
 
       return completer.future;
     });
+  }
+}
+
+/// A wrapper around the `test` package. This class is used to run your unit
+/// tests.
+class TestRunner {
+  final PubApp _test = new PubApp.local('test');
+
+  TestRunner();
+
+  /// Run the tests in the current package. See the
+  /// [test package](https://pub.dartlang.org/packages/test).
+  ///
+  /// [name] is substring of the name of the test to run. Regular expression
+  /// syntax is supported. [plainName] is a plain-text substring of the name of
+  /// the test to run. [platformSelector] is the platform(s) on which to run the
+  /// tests. This parameter can be a String or a List.
+  /// [Available values](https://github.com/dart-lang/test#platform-selector-syntax)
+  /// are `vm` (default), `dartium`, `content-shell`, `chrome`, `phantomjs`,
+  /// `firefox`, `safari`. [concurrency] controls the number of concurrent test
+  /// suites run (defaults to 4). [pubServe] is the port of a pub serve instance
+  /// serving `test/`.
+  void test({String name, String plainName, dynamic platformSelector,
+      int concurrency, int pubServe, RunOptions runOptions}) {
+    _test.run(_buildArgs(
+        name: name,
+        plainName: plainName,
+        selector: platformSelector,
+        concurrency: concurrency,
+        pubServe: pubServe), script: 'test', runOptions: runOptions);
+  }
+
+  /// Run the tests in the current package. See the
+  /// [test package](https://pub.dartlang.org/packages/test).
+  ///
+  /// [name] is substring of the name of the test to run. Regular expression
+  /// syntax is supported. [plainName] is a plain-text substring of the name of
+  /// the test to run. [platformSelector] is the platform(s) on which to run the
+  /// tests. This parameter can be a String or a List.
+  /// [Available values](https://github.com/dart-lang/test#platform-selector-syntax)
+  /// are `vm` (default), `dartium`, `content-shell`, `chrome`, `phantomjs`,
+  /// `firefox`, `safari`. [concurrency] controls the number of concurrent test
+  /// suites run (defaults to 4). [pubServe] is the port of a pub serve instance
+  /// serving `test/`.
+  Future testAsync({String name, String plainName, dynamic platformSelector,
+      int concurrency, int pubServe, RunOptions runOptions}) {
+    return _test.runAsync(_buildArgs(
+        name: name,
+        plainName: plainName,
+        selector: platformSelector,
+        concurrency: concurrency,
+        pubServe: pubServe), script: 'test', runOptions: runOptions);
+  }
+
+  List<String> _buildArgs({String name, String plainName, dynamic selector,
+      int concurrency, int pubServe}) {
+    List<String> args = ['--reporter=expanded'];
+    if (name != null) args.add('--name=${name}');
+    if (plainName != null) args.add('--plain-name=${plainName}');
+    if (selector != null) {
+      if (selector is List) selector = selector.join('||');
+      args.add('--platform=${selector}');
+    }
+    if (concurrency != null) args.add('--concurrency=${concurrency}');
+    if (pubServe != null) args.add('--pub-serve=${pubServe}');
+    // TODO: Pass in --color based on a global property: #243.
+    return args;
   }
 }