+++ changelog.md
@@ -11,6 +11,7 @@
 - Added support for passing lists of files or directories to `DartFmt.format()`.
 - `Analyzer.analyzeFiles` is deprecated in favor of `Analyzer.analyze`.
 - Added a `Pub.global.list()` method to list the installed applications.
+- added a `TestRunner` class - a wrapper around the new `test` package.
 
 ## 0.7.0 (2015/4/20)
 - Big changes! Task definitions are now primarily annotation based; see the